@@ -2,67 +2,70 @@
   "name": "handle-mini",
   "version": "1.0.0",
   "private": true,
-  "packageManager": "pnpm@9.1.2",
+  "type": "module",
+  "packageManager": "pnpm@9.1.3",
   "scripts": {
-    "lint:script": "cross-env NODE_ENV=production eslint --ext .js,.ts .",
-    "lint:style": "cross-env NODE_ENV=production stylelint \"src/**/*.less\"",
+    "format": "prettier --write \"**/*.{js,cjs,ts,json,less,html,md}\"",
+    "lint:script": "eslint --ext .js,.cjs,.ts .",
+    "lint:style": "stylelint \"src/**/*.less\"",
     "type": "tsc",
-    "start": "cross-env NODE_ENV=development node build.js",
+    "dev": "cross-env NODE_ENV=development node build.js",
     "build": "cross-env NODE_ENV=production node build.js",
-    "prepare": "husky install"
+    "prepare": "husky"
   },
   "dependencies": {
-    "@babel/runtime": "^7.20.7",
+    "@babel/runtime": "^7.24.6",
     "@hankit/tools": "workspace:*",
-    "@vue-mini/wechat": "^0.3.0",
+    "@vue-mini/core": "1.0.0-rc.11",
     "seedrandom": "^3.0.5"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.7",
+    "@babel/core": "^7.24.6",
     "@babel/plugin-transform-nullish-coalescing-operator": "^7.24.6",
     "@babel/plugin-transform-numeric-separator": "^7.24.6",
     "@babel/plugin-transform-optional-chaining": "^7.24.6",
-    "@babel/plugin-transform-runtime": "^7.19.6",
-    "@babel/preset-env": "^7.20.2",
-    "@babel/preset-typescript": "^7.18.6",
-    "@rollup/plugin-node-resolve": "^15.0.1",
-    "@rollup/plugin-replace": "^5.0.2",
-    "@rollup/plugin-terser": "^0.2.1",
-    "@types/pinyin": "^2.10.0",
-    "@types/seedrandom": "^3.0.3",
-    "@typescript-eslint/eslint-plugin": "^5.47.0",
-    "@typescript-eslint/parser": "^5.47.0",
+    "@babel/plugin-transform-runtime": "^7.24.6",
+    "@babel/preset-env": "^7.24.6",
+    "@babel/preset-typescript": "^7.24.6",
+    "@babel/traverse": "^7.24.6",
+    "@babel/types": "^7.24.6",
+    "@rollup/plugin-commonjs": "^25.0.8",
+    "@rollup/plugin-node-resolve": "^15.2.3",
+    "@rollup/plugin-replace": "^5.0.5",
+    "@rollup/plugin-terser": "^0.4.4",
+    "@types/pinyin": "^2.10.2",
+    "@types/seedrandom": "^3.0.8",
     "babel-plugin-autocomplete-index": "^0.2.0",
-    "babel-plugin-module-resolver": "^4.1.0",
+    "babel-plugin-module-resolver": "^5.0.2",
     "babel-plugin-transform-inline-environment-variables": "^0.4.4",
-    "chokidar": "^3.5.3",
+    "chokidar": "^3.6.0",
     "cross-env": "^7.0.3",
-    "eslint": "^8.30.0",
-    "eslint-config-silent": "^0.30.0",
-    "eslint-config-xo-typescript": "^0.55.1",
-    "eslint-import-resolver-typescript": "^3.5.2",
-    "fs-extra": "^11.1.0",
-    "husky": "^8.0.2",
-    "less": "^4.1.3",
-    "lint-staged": "^13.1.0",
-    "postcss": "^8.4.20",
+    "eslint-import-resolver-typescript": "^3.6.1",
+    "fs-extra": "^11.2.0",
+    "husky": "^9.0.11",
+    "less": "^4.2.0",
+    "lint-staged": "^15.2.5",
+    "postcss": "^8.4.38",
     "postcss-less": "^6.0.0",
     "postcss-pxtorpx-pro": "^2.0.0",
-    "postcss-url": "^10.1.3",
-    "posthtml": "^0.16.6",
-    "prettier": "^2.8.1",
-    "rollup": "^3.8.1",
-    "serve": "^14.1.2",
-    "stylelint": "^14.16.0",
-    "stylelint-config-prettier": "^9.0.4",
-    "stylelint-config-xo": "^0.21.1",
-    "stylelint-prettier": "^2.0.0",
-    "terser": "^5.16.1",
-    "typescript": "^4.9.4",
-    "xo": "^0.53.1"
+    "rollup": "^4.18.0",
+    "stylelint": "^16.6.0",
+    "stylelint-config-standard": "^36.0.0",
+    "terser": "^5.31.0",
+    "typescript": "^5.4.5",
+    "xo": "^0.58.0"
+  },
+  "pnpm": {
+    "peerDependencyRules": {
+      "ignoreMissing": [
+        "eslint",
+        "eslint-plugin-import",
+        "webpack"
+      ]
+    }
   },
   "browserslist": [
     "iOS >= 14",
-    "Chrome >= 86"
+    "Chrome >= 116"
   ]
 }
