@@ -1,93 +1,37 @@
-/* eslint-disable unicorn/prefer-module */
-'use strict';
-
-const os = require('node:os');
-const path = require('node:path');
-const crypto = require('node:crypto');
-const process = require('node:process');
-const { spawn } = require('node:child_process');
-const fs = require('fs-extra');
-const chokidar = require('chokidar');
-const babel = require('@babel/core');
-const { minify } = require('terser');
-const posthtml = require('posthtml');
-const less = require('less');
-const postcss = require('postcss');
-const pxtorpx = require('postcss-pxtorpx-pro');
-const url = require('postcss-url');
-const { rollup } = require('rollup');
-const replace = require('@rollup/plugin-replace');
-const terser = require('@rollup/plugin-terser');
-const { default: resolve } = require('@rollup/plugin-node-resolve');
+import path from 'node:path';
+import process from 'node:process';
+import fs from 'fs-extra';
+import chokidar from 'chokidar';
+import babel from '@babel/core';
+import traverse from '@babel/traverse';
+import t from '@babel/types';
+import { minify } from 'terser';
+import less from 'less';
+import postcss from 'postcss';
+import pxtorpx from 'postcss-pxtorpx-pro';
+import { rollup } from 'rollup';
+import replace from '@rollup/plugin-replace';
+import terser from '@rollup/plugin-terser';
+import resolve from '@rollup/plugin-node-resolve';
+import commonjs from '@rollup/plugin-commonjs';
 
 const NODE_ENV = process.env.NODE_ENV || 'production';
 const __PROD__ = NODE_ENV === 'production';
-const localPath = `http://${getLocalIP()}:3000/`;
-const publicPath = 'https://your.static.server/';
-const terserOptions = { ecma: 2015, toplevel: true, safari10: true };
-
-process.on('unhandledRejection', (error) => {
-  throw error;
-});
-
-function getLocalIP() {
-  const ifaces = Object.values(os.networkInterfaces());
-  for (const iface of ifaces) {
-    for (const alias of iface) {
-      if (alias.internal || alias.family !== 'IPv4') continue;
-      return alias.address;
-    }
-  }
-}
-
-function getImagePathWithHash(imagePath, originPath) {
-  const buffer = fs.readFileSync(imagePath);
-  const hash = crypto
-    .createHash('md5')
-    .update(buffer)
-    .digest('hex')
-    .slice(0, 8);
-  return originPath.replace(/\.(png|jpe?g|gif|svg)$/, `.${hash}.$1`);
-}
+const terserOptions = {
+  ecma: 2016,
+  toplevel: true,
+  format: { comments: false },
+};
 
 const bundledModules = new Set();
 async function bundleModule(module) {
   if (bundledModules.has(module)) return;
   bundledModules.add(module);
 
-  if (module === 'seedrandom') {
-    const filePath = require.resolve(`${module}/seedrandom.min.js`);
-    const destination = `dist/miniprogram_npm/${module}/index.js`;
-    fs.copy(filePath, destination);
-    return;
-  }
-
-  let pkg;
-  try {
-    pkg = require(`${module}/package.json`);
-  } catch {}
-
-  if (pkg && !pkg.module) {
-    throw new Error(`Can't found esm bundle of ${module}`);
-  }
-
-  let entry = pkg
-    ? path.join(
-        path.dirname(require.resolve(`${module}/package.json`)),
-        pkg.module
-      )
-    : require.resolve(module);
-
-  if (module.startsWith('@babel/runtime/')) {
-    const paths = entry.split(path.sep);
-    paths.splice(-1, 0, 'esm');
-    // Path.join 在 POSIX OS 上会返回 'root/**' 而非正确的 '/root/**'，所以不能使用。
-    entry = paths.join(path.sep);
-  }
-
   const bundle = await rollup({
-    input: entry,
+    input: module,
     plugins: [
+      commonjs(),
       replace({
         preventAssignment: true,
         values: {
@@ -106,11 +50,23 @@ async function bundleModule(module) {
 }
 
 async function processScript(filePath) {
-  let { code } = await babel.transformFileAsync(path.resolve(filePath));
-  for (const [, module] of code.matchAll(/require\("(.+?)"\)/g)) {
-    if (module.startsWith('.')) continue;
-    bundleModule(module);
-  }
+  let { ast, code } = await babel.transformFileAsync(path.resolve(filePath), {
+    ast: true,
+  });
+
+  traverse.default(ast, {
+    CallExpression({ node }) {
+      if (
+        node.callee.name !== 'require' ||
+        !t.isStringLiteral(node.arguments[0]) ||
+        node.arguments[0].value.startsWith('.')
+      ) {
+        return;
+      }
+
+      bundleModule(node.arguments[0].value);
+    },
+  });
 
   if (__PROD__) {
     // eslint-disable-next-line unicorn/no-await-expression-member
@@ -123,40 +79,14 @@ async function processScript(filePath) {
   fs.writeFile(destination, code);
 }
 
-async function processTemplate(filePath, origin = localPath) {
-  const source = await fs.readFile(filePath, 'utf8');
-  const { html } = await posthtml()
-    .use((tree) => {
-      tree.match([{ tag: 'image' }, { tag: 'cover-image' }], (node) => {
-        const { src } = node.attrs;
-        if (
-          /^{{.*}}$/.test(src) ||
-          /^https?:\/\//.test(src) ||
-          src.startsWith('data:') ||
-          src.includes('/assets/')
-        ) {
-          return node;
-        }
-
-        const absolutePath = src.startsWith('/')
-          ? path.resolve('src', src.slice(1))
-          : path.resolve(path.dirname(filePath), src);
-        const href =
-          origin + path.relative('src', absolutePath).replace(/\\/g, '/');
-        node.attrs.src = __PROD__
-          ? getImagePathWithHash(absolutePath, href)
-          : href;
-        return node;
-      });
-    })
-    .process(source, { xmlMode: true, singleTags: [/<>/] });
-  const destination = filePath.replace('src', 'dist');
-  // Make sure the directory already exists when write file
+async function processTemplate(filePath) {
+  const destination = filePath
+    .replace('src', 'dist')
+    .replace(/\.html$/, '.wxml');
   await fs.copy(filePath, destination);
-  fs.writeFile(destination, html);
 }
 
-async function processStyle(filePath, origin = localPath) {
+async function processStyle(filePath) {
   let source = await fs.readFile(filePath, 'utf8');
   source =
     `@import '${path.resolve('src/styles/variables.less')}';\n` +
@@ -167,22 +97,6 @@ async function processStyle(filePath, origin = localPath) {
   });
   const { css: wxss } = await postcss()
     .use(pxtorpx({ minPixelValue: 2 }))
-    .use(
-      url({
-        url(asset) {
-          if (/^https?:\/\//.test(asset.url) || asset.url.startsWith('data:')) {
-            return asset.url;
-          }
-
-          const absolutePath = asset.url.startsWith('/')
-            ? path.resolve('src', asset.url.slice(1))
-            : path.resolve(path.dirname(filePath), asset.url);
-          const href =
-            origin + path.relative('src', absolutePath).replace(/\\/g, '/');
-          return __PROD__ ? getImagePathWithHash(absolutePath, href) : href;
-        },
-      })
-    )
     .process(css, { map: false, from: undefined });
   const destination = filePath
     .replace('src', 'dist')
@@ -208,7 +122,7 @@ async function dev() {
       return;
     }
 
-    if (/\.wxml$/.test(filePath)) {
+    if (/\.html$/.test(filePath)) {
       processTemplate(filePath);
       return;
     }
@@ -241,7 +155,6 @@ async function dev() {
 
 async function prod() {
   await fs.remove('dist');
-  await fs.remove('temp');
   const watcher = chokidar.watch(['src', '!src/styles/**/*'], {
     ignored: ['**/.{gitkeep,DS_Store}'],
   });
@@ -251,21 +164,13 @@ async function prod() {
       return;
     }
 
-    if (/\.wxml$/.test(filePath)) {
-      processTemplate(filePath, publicPath);
+    if (/\.html$/.test(filePath)) {
+      processTemplate(filePath);
       return;
     }
 
     if (/\.less$/.test(filePath)) {
-      processStyle(filePath, publicPath);
-      return;
-    }
-
-    if (filePath.includes(path.join('src', 'images'))) {
-      fs.copy(
-        filePath,
-        getImagePathWithHash(filePath, filePath.replace('src', 'temp'))
-      );
+      processStyle(filePath);
       return;
     }
 
@@ -274,11 +179,9 @@ async function prod() {
   watcher.on('ready', () => watcher.close());
 }
 
+// eslint-disable-next-line unicorn/prefer-ternary
 if (__PROD__) {
-  // eslint-disable-next-line unicorn/prefer-top-level-await
-  prod();
+  await prod();
 } else {
-  spawn('serve', ['src'], { stdio: 'inherit', shell: true });
-  // eslint-disable-next-line unicorn/prefer-top-level-await
-  dev();
+  await dev();
 }
