@@ -6,7 +6,7 @@ import { idioms } from '@/data/idioms';
 const getPinyinRawWithTypes = getPinyinRaw as typeof import('pinyin');
 
 export function getIdiom(
-  word: string
+  word: string,
 ): [string, string | undefined] | undefined {
   const simplified = toSimplified(word);
   if (polyphones[word]) return [word, polyphones[word]];
@@ -18,11 +18,12 @@ export function getIdiom(
 
 export function getPinyin(word: string) {
   const data = getIdiom(word);
-  const parts = data?.[1]
-    ? data[1].split(/\s+/g)
+  const parts =
+    data?.[1] ?
+      data[1].split(/\s+/g)
     : getPinyinRawWithTypes(data?.[0] ?? toSimplified(word), {
         style: getPinyinRawWithTypes.STYLE_TONE2,
       }).map((i) => i[0]);
   // https://baike.baidu.com/item/%E6%B1%89%E8%AF%AD%E6%8B%BC%E9%9F%B3%E6%96%B9%E6%A1%88/1884432
-  return parts.map((i) => i.replace(/^([jqxy])u([a-z]*\d?)$/g, '$1v$2'));
+  return parts.map((i) => i.replaceAll(/^([jqxy])u([a-z]*\d?)$/g, '$1v$2'));
 }
